[Snapshot] Split implementation and supplemental.
This change splits out the implementation-specific files (required for
writing a Snapshot test case) from the supplemental files (conveniences
for making the test cases easier to write). This work is part of #7253.

Closes #7354
