+++ components/private/Snapshot/src/MaterialSnapshot.h
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "MDCSnapshotTestCase.h"
-#import "UIImage+MDCSnapshot.h"
-#import "UIView+MDCSnapshot.h"
+#import "Implementation/MDCSnapshotTestCase.h"
+#import "Supplemental/UIImage+MDCSnapshot.h"
+#import "Supplemental/UIView+MDCSnapshot.h"
