+++ components/private/Snapshot/src/Supplemental/UIImage+MDCSnapshot.h
