+++ components/private/Snapshot/src/Supplemental/UIImage+MDCSnapshot.m
