+++ components/private/Snapshot/src/Supplemental/UIView+MDCSnapshot.h
